refactor: Convert to Python package
This commit refactors the OASIS codebase into a proper Python package. This improves code organization, maintainability, and allows for easier distribution and installation. The main functionality remains unchanged.
