+++ oasis/__init__.py
@@ -1,0 +1,7 @@
+"""
+OASIS - Ollama Automated Security Intelligence Scanner
+"""
+
+from .oasis import main
+
+__version__ = "0.2.0"
